game engine works. made and tested code for encoders, dc motors, servos, and proporitonal distance control using dc motor + encoder

+++ Final/Final0/MyMath.h
@@ -3,7 +3,7 @@
 class MyMath
 {
   public:
-  static int sign(int var)
+  static float sign(float var)
   {
     if (var == 0)
     {
